requirements.txt
fastapi
uvicorn
opencv-python
pygame
Pillow
websockets
mediapipe